fixed migration files

+++ db/migrate/20210912015647_create_notifications.rb
@@ -1,10 +1,10 @@
 class CreateNotifications < ActiveRecord::Migration[5.2]
   def change
     create_table :notifications do |t|
-      t.integer :visiter_id, null: false
-      t.integer :visited_id, null: false
-      t.integer :post_id
-      t.integer :post_comment_id
+      t.bigint :visiter_id, null: false
+      t.bigint :visited_id, null: false
+      t.bigint :post_id
+      t.bigint :post_comment_id
       t.string :action, default: "", null: false
       t.boolean :checked, default: false, null: false
       t.timestamps
