+++ db/migrate/20210915074911_create_login_histories.rb
@@ -1,7 +1,7 @@
 class CreateLoginHistories < ActiveRecord::Migration[5.2]
   def change
     create_table :login_histories do |t|
-      t.integer :user_id, null: false
+      t.bigint :user_id, null: false
       t.datetime :logind_at, null: false
 
       t.timestamps
